renamed the file containing global_config to global_config.py
