+++ inputremapper/injection/mapping_handlers/joystick_to_mouse.py
@@ -39,7 +39,7 @@ from evdev.ecodes import (
 )
 
 from inputremapper.logger import logger
-from inputremapper.config import MOUSE, WHEEL
+from inputremapper.global_config import MOUSE, WHEEL
 from inputremapper import utils
 from inputremapper.injection.mapping_handlers.consumer import Consumer
 from inputremapper.groups import classify, GAMEPAD
