+++ tests/integration/test_gui.py
@@ -49,7 +49,7 @@ from gi.repository import Gtk, GLib, Gdk, GtkSource
 from inputremapper.system_mapping import system_mapping, XMODMAP_FILENAME
 from inputremapper.gui.active_preset import active_preset
 from inputremapper.paths import CONFIG_PATH, get_preset_path, get_config_path
-from inputremapper.config import global_config, WHEEL, MOUSE, BUTTONS
+from inputremapper.global_config import global_config, WHEEL, MOUSE, BUTTONS
 from inputremapper.gui.reader import reader
 from inputremapper.gui.helper import RootHelper
 from inputremapper.gui.utils import gtk_iteration
