+++ tests/unit/test_config.py
@@ -22,7 +22,7 @@
 import os
 import unittest
 
-from inputremapper.config import global_config
+from inputremapper.global_config import global_config
 from inputremapper.paths import touch
 
 from tests.test import quick_cleanup, tmp
