+++ tests/unit/test_injector.py
@@ -61,7 +61,7 @@ from inputremapper.injection.numlock import is_numlock_on, set_numlock, ensure_n
 from inputremapper.system_mapping import system_mapping, DISABLE_CODE, DISABLE_NAME
 from inputremapper.gui.active_preset import active_preset
 from inputremapper.preset import Preset
-from inputremapper.config import global_config, NONE, MOUSE, WHEEL, BUTTONS
+from inputremapper.global_config import global_config, NONE, MOUSE, WHEEL, BUTTONS
 from inputremapper.key import Key
 from inputremapper.injection.macros.parse import parse
 from inputremapper.injection.context import Context
